+++ src/api.ts
@@ -1,4 +1,4 @@
-import { getScopeComponent, addMany as addManyInternal } from './api/consumer/index';
+import { getScopeComponent, addMany as addManyInternal, build, buildAll as buildAllApi } from './api/consumer/index';
 import { AddProps } from './consumer/component-ops/add-components/add-components';
 import { scopeList } from './api/scope/index';
 import Extension from './extensions/extension';
@@ -19,6 +19,24 @@ export function show(scopePath: string, id: string, opts?: Record<string, any>) 
 export function list(scopePath: string) {
   return scopeList(scopePath).then(listScopeResult => listScopeResult.map(result => result.id.toString()));
 }
+
+export async function buildOne(
+  id: string,
+  noCache = false,
+  verbose = false,
+  workspaceDir?: string
+): Promise<string[] | undefined> {
+  return build(id, noCache, verbose, workspaceDir);
+}
+
+export async function buildAll(
+  id: string,
+  noCache = false,
+  verbose = false,
+): Promise<Record<string, any>> {
+  return buildAllApi(noCache, verbose);
+}
+
 export async function addMany(components: AddProps[], alternateCwd?: string) {
   return addManyInternal(components, alternateCwd);
 }