+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [14.4.4-dev.8] - 2019-11-21
+
+- refactor buildOne to include workingDir
+
+## [14.4.4-dev.6] - 2019-11-21
+
+- fix symlink errors when generating links to `d.ts` files
+- expose buildOne and buildAll for programmatic api (Experimental)
+
+## [14.4.4-dev.5] - 2019-11-19
+
+- [#2140](https://github.com/teambit/bit/issues/2140) update bit-javascript to support `import { x as y }` syntax
+
 ## [14.4.4-dev.4] - 2019-11-18
 
 - [#2132](https://github.com/teambit/bit/issues/2132) fix compiler import to install peer dependencies